Average Loss

Calculates average loss, optionally grouped by a discrete vector. The function supports multivariate losses and case weights.

average_loss(object, ...)

# S3 method for default
average_loss(
  object,
  X,
  y,
  pred_fun = stats::predict,
  BY = NULL,
  loss = "squared_error",
  w = NULL,
  ...
)

# S3 method for ranger
average_loss(
  object,
  X,
  y,
  pred_fun = function(m, X, ...) stats::predict(m, X, ...)$predictions,
  BY = NULL,
  loss = "squared_error",
  w = NULL,
  ...
)

# S3 method for Learner
average_loss(
  object,
  v,
  X,
  y,
  pred_fun = NULL,
  BY = NULL,
  loss = "squared_error",
  w = NULL,
  ...
)

# S3 method for explainer
average_loss(
  object,
  X = object[["data"]],
  y = object[["y"]],
  pred_fun = object[["predict_function"]],
  BY = NULL,
  loss = "squared_error",
  w = object[["weights"]],
  ...
)

Arguments

object

Fitted model object.

...

Additional arguments passed to pred_fun(object, X, ...), for instance type = "response" in a glm() model.

X

A data.frame or matrix serving as background dataset.

y

Numeric vector or matrix of the response corresponding to X.

pred_fun

Prediction function of the form function(object, X, ...), providing \(K \ge 1\) numeric predictions per row. Its first argument represents the model object, its second argument a data structure like X. Additional arguments (such as type = "response" in a GLM) can be passed via .... The default, stats::predict(), will work in most cases. Note that column names in a resulting matrix of predictions will be used as default column names in the results.

BY

Optional grouping vector.

loss

One of "squared_error", "logloss", "mlogloss", "poisson", "gamma", "absolute_error", or a loss function that turns observed and predicted values (vectors or matrices) into a vector or matrix of unit losses. For "mlogloss", the response y can either be a matrix with one column per category or a vector with categories. The latter case is internally exploded to the shape of the predictions via stats::model.matrix(~ y + 0).

w

Optional vector of case weights for each row of X.

v

Vector of feature names.

Value

A matrix with one row per group and one column per loss dimension.

Methods (by class)

average_loss(default): Default method.

average_loss(ranger): Method for "ranger" models.

average_loss(Learner): Method for "mlr3" models.

average_loss(explainer): Method for DALEX "explainer".

Examples

# MODEL 1: Linear regression
fit <- lm(Sepal.Length ~ ., data = iris)
average_loss(fit, X = iris, y = iris$Sepal.Length)
#>            [,1]
#> [1,] 0.09037657
average_loss(fit, X = iris, y = iris$Sepal.Length, BY = iris$Species)
#>                  [,1]
#> setosa     0.06463433
#> versicolor 0.10839144
#> virginica  0.09810394

# MODEL 2: Multi-response linear regression
fit <- lm(as.matrix(iris[1:2]) ~ Petal.Length + Petal.Width + Species, data = iris)
average_loss(fit, X = iris, y = iris[1:2])
#>      Sepal.Length Sepal.Width
#> [1,]    0.1112099  0.08472089
average_loss(fit, X = iris, y = iris[1:2], loss = "gamma", BY = iris$Species)
#>            Sepal.Length Sepal.Width
#> setosa      0.004646018 0.011500586
#> versicolor  0.003121888 0.007489254
#> virginica   0.002525590 0.007419552