Summary Method

Summary method for "perm_importance" object. By default shows the top 15 feature importances collapsed over repetitions (and loss dimensions).

# S3 method for perm_importance
summary(
  object,
  normalize = FALSE,
  sort = TRUE,
  top_m = 15L,
  agg_cols = c("sum", "mean", "max", "no"),
  err_type = c("se", "sd", "no"),
  ...
)

Arguments

object

An object of class "perm_importance".

normalize

Should importance statistics be divided by performance? Default is FALSE. Normalization is applied after collapsing multivariate losses.

sort

Should results be sorted by importance? Default is TRUE. The multivariate case is sorted by row sums.

top_m

Of how many features should importance values be shown? Default is Inf, i.e., show all.

agg_cols

Name of function used to collapse multivariate losses. Default is "sum". Set to "no" for no aggregation. Other options are "mean" and "max". This step is applied before normalization.

err_type

Should standard errors ("se", default) or standard deviations ("sd") be calculated? Only relevant if perms > 1. Set to "no" to hide errors.

...

Currently not used.

Value

A list with two matrices: imp contains the importance values per feature, while err represents the uncertainty of these values.

See also